Mapping through routes was causing the render to flash. Unsure why but moving back to a static component for initial release.
