+++ src/pages/AccountSettings/AccountSettings.js
@@ -1,12 +1,19 @@
-import { Suspense } from 'react'
+import { Suspense, lazy } from 'react'
 import { useParams, Switch, Route, Redirect } from 'react-router-dom'
-import uniqueId from 'lodash/uniqueId'
 
 import LogoSpinner from 'ui/LogoSpinner'
 import SidebarLayout from 'layouts/SidebarLayout'
 
 import SideMenu from './SideMenu'
-import routes from './routes'
+
+const CancelPlanTab = lazy(() => import('./tabs/CancelPlan'))
+const UpgradePlanTab = lazy(() => import('./tabs/UpgradePlan'))
+const InvoicesTab = lazy(() => import('./tabs/Invoices'))
+const InvoiceDetailTab = lazy(() => import('./tabs/InvoiceDetail'))
+const BillingAndUsersTab = lazy(() => import('./tabs/BillingAndUsers'))
+const AdminTab = lazy(() => import('./tabs/Admin'))
+const YAMLTab = lazy(() => import('./tabs/YAML'))
+const AccessTab = lazy(() => import('./tabs/Access'))
 
 function AccountSettings() {
   const { provider, owner } = useParams()
@@ -21,14 +28,33 @@ function AccountSettings() {
     <SidebarLayout sidebar={<SideMenu />}>
       <Suspense fallback={tabLoading}>
         <Switch>
-          {routes.map(({ path, exact, Component, redirect }) => (
-            <Route key={uniqueId(path)} path={path} exact={exact}>
-              {redirect && (
-                <Redirect path={path} to={redirect({ provider, owner })} />
-              )}
-              {Component && <Component provider={provider} owner={owner} />}
-            </Route>
-          ))}
+          <Route path="/account/:provider/:owner/" exact>
+            <AdminTab provider={provider} owner={owner} />
+          </Route>
+          <Route path="/account/:provider/:owner/yaml/" exact>
+            <YAMLTab />
+          </Route>
+          <Route path="/account/:provider/:owner/access/" exact>
+            <AccessTab />
+          </Route>
+          <Route path="/account/:provider/:owner/billing/" exact>
+            <BillingAndUsersTab provider={provider} owner={owner} />
+          </Route>
+          <Route path="/account/:provider/:owner/users/" exact>
+            <Redirect to={`/account/${provider}/${owner}/billing/`} />
+          </Route>
+          <Route path="/account/:provider/:owner/billing/upgrade/" exact>
+            <UpgradePlanTab provider={provider} owner={owner} />
+          </Route>
+          <Route path="/account/:provider/:owner/billing/cancel/" exact>
+            <CancelPlanTab provider={provider} owner={owner} />
+          </Route>
+          <Route path="/account/:provider/:owner/invoices/" exact>
+            <InvoicesTab provider={provider} owner={owner} />
+          </Route>
+          <Route path="/account/:provider/:owner/invoices/:id/" exact>
+            <InvoiceDetailTab provider={provider} owner={owner} />
+          </Route>
         </Switch>
       </Suspense>
     </SidebarLayout>
